refactor: Use custom database tables for authentication and authorization
- Replaced Spring Security's default tables with custom tables for users and roles.
- Updated database schema to define and configure the new custom tables.
- Refactored authentication and authorization logic to integrate with custom tables.
- Modified user and role queries to match the new table structure.
- Tested the updated functionality thoroughly using Postman to ensure correctness.
- Updated documentation to reflect the new table structure and code changes.
