+++ .idea/workspace.xml
@@ -4,8 +4,15 @@
     <option name="autoReloadType" value="SELECTIVE" />
   </component>
   <component name="ChangeListManager">
-    <list default="true" id="1de4406e-2fa5-4f6b-b82d-cfe32e97160e" name="Changes" comment="feat: Add JDBC-based authorization using users and authorities tables&#10;&#10;- Set up 'users' table to store user credentials (username, password, enabled).&#10;- Set up 'authorities' table to store user roles/permissions with foreign key referencing 'users' table.&#10;- Integrated JDBC connector to handle database-based authorization.&#10;- Configured database schema for referential integrity between 'users' and 'authorities'.&#10;- Ensured roles and credentials are loaded dynamically from the database for authorization.&#10;- Added functionality to enforce role-based access control (RBAC) via JDBC authentication.">
+    <list default="true" id="1de4406e-2fa5-4f6b-b82d-cfe32e97160e" name="Changes" comment="chore: Implement bcrypt hashing for passwords and update authentication logic">
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/com/luv2code/springboot/cruddemo/dao/EmployeeRepository.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/com/luv2code/springboot/cruddemo/dao/EmployeeRepository.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/com/luv2code/springboot/cruddemo/entity/Employee.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/com/luv2code/springboot/cruddemo/rest/entity/Employee.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/com/luv2code/springboot/cruddemo/rest/EmployeeRestController.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/com/luv2code/springboot/cruddemo/rest/EmployeeRestController.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/com/luv2code/springboot/cruddemo/security/DemoSecurityConfig.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/com/luv2code/springboot/cruddemo/security/DemoSecurityConfig.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/com/luv2code/springboot/cruddemo/service/EmployeeService.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/com/luv2code/springboot/cruddemo/service/EmployeeService.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/com/luv2code/springboot/cruddemo/service/EmployeeServiceImpl.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/com/luv2code/springboot/cruddemo/service/EmployeeServiceImpl.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/resources/application.properties" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/resources/application.properties" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -153,7 +160,14 @@
       <option name="project" value="LOCAL" />
       <updated>1735070267234</updated>
     </task>
-    <option name="localTasksCounter" value="11" />
+    <task id="LOCAL-00011" summary="chore: Implement bcrypt hashing for passwords and update authentication logic">
+      <created>1735073209833</created>
+      <option name="number" value="00011" />
+      <option name="presentableId" value="LOCAL-00011" />
+      <option name="project" value="LOCAL" />
+      <updated>1735073209833</updated>
+    </task>
+    <option name="localTasksCounter" value="12" />
     <servers />
   </component>
   <component name="Vcs.Log.Tabs.Properties">
@@ -178,6 +192,7 @@
     <MESSAGE value="feat: Add in-memory user details with role-based access control&#10;&#10;- Configured Spring Security to use in-memory user details.&#10;- Added three users with different roles: ADMIN, MANAGER, and EMPLOYEE.&#10;- Implemented role-based access control for API endpoints.&#10;- Tested endpoint accessibility based on user roles to ensure proper security." />
     <MESSAGE value="feat: Add security filter chain with role-based access control&#10;&#10;- Implemented filterChain method to configure HTTP security with role-based access control.&#10;- Restricted access to API endpoints based on roles: &quot;EMPLOYEE&quot;, &quot;MANAGER&quot;, and &quot;ADMIN&quot;.&#10;- Disabled CSRF protection and enabled HTTP Basic authentication." />
     <MESSAGE value="feat: Add JDBC-based authorization using users and authorities tables&#10;&#10;- Set up 'users' table to store user credentials (username, password, enabled).&#10;- Set up 'authorities' table to store user roles/permissions with foreign key referencing 'users' table.&#10;- Integrated JDBC connector to handle database-based authorization.&#10;- Configured database schema for referential integrity between 'users' and 'authorities'.&#10;- Ensured roles and credentials are loaded dynamically from the database for authorization.&#10;- Added functionality to enforce role-based access control (RBAC) via JDBC authentication." />
-    <option name="LAST_COMMIT_MESSAGE" value="feat: Add JDBC-based authorization using users and authorities tables&#10;&#10;- Set up 'users' table to store user credentials (username, password, enabled).&#10;- Set up 'authorities' table to store user roles/permissions with foreign key referencing 'users' table.&#10;- Integrated JDBC connector to handle database-based authorization.&#10;- Configured database schema for referential integrity between 'users' and 'authorities'.&#10;- Ensured roles and credentials are loaded dynamically from the database for authorization.&#10;- Added functionality to enforce role-based access control (RBAC) via JDBC authentication." />
+    <MESSAGE value="chore: Implement bcrypt hashing for passwords and update authentication logic" />
+    <option name="LAST_COMMIT_MESSAGE" value="chore: Implement bcrypt hashing for passwords and update authentication logic" />
   </component>
 </project>
