+++ src/main/java/com/luv2code/springboot/cruddemo/dao/EmployeeRepository.java
@@ -1,6 +1,6 @@
 package com.luv2code.springboot.cruddemo.dao;
 
-import com.luv2code.springboot.cruddemo.entity.Employee;
+import com.luv2code.springboot.cruddemo.rest.entity.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface EmployeeRepository extends JpaRepository<Employee, Integer> {
