+++ src/main/java/com/luv2code/springboot/cruddemo/rest/entity/Employee.java
@@ -1,4 +1,4 @@
-package com.luv2code.springboot.cruddemo.entity;
+package com.luv2code.springboot.cruddemo.rest.entity;
 
 import jakarta.persistence.*;
