+++ src/main/java/com/luv2code/springboot/cruddemo/service/EmployeeService.java
@@ -1,6 +1,6 @@
 package com.luv2code.springboot.cruddemo.service;
 
-import com.luv2code.springboot.cruddemo.entity.Employee;
+import com.luv2code.springboot.cruddemo.rest.entity.Employee;
 
 import java.util.List;
