+++ src/main/resources/application.properties
@@ -10,3 +10,4 @@ spring.security.user.password = test123
 
 
 server.port = 8081
+logging.level.org.springframework.security=DEBUG
